Add safe user insertion to api; refactor
Added post routing to /users to handle insertion requests to routes.js; right now this
only supports 201 (success) and 404 (not found) codes; code 409 (conflict) needs to
be added. The response code also comes with an object indicating success or failure,
along with a brief message.

A 'safe query' function was added to the database interface, so that the insertion
input is sanitized.

postHandlers.js modified so that input parameters from request are able to be read using
bodyParser.

The user controller now contains a creation method which takes all parameters which
compose a user. Originally, I wanted to use a user instance as input, but passing
each component seems like a simple route for the time-being, but will need to be
evaluated.