Add random user pair mechanism
Add table to hold the randomization results. This table has
not primary key, as it is more of a data store, and so such
constraints are not necessary. We still want to reference the
two other involved tables: events and users, so constraints
are placed on those fields. This table was also added to the
dropAllTables.sql script, as this new table must be initialized
along with the other tables.

Add routines to grab the users for each event from the database
and create a list of [giver_id, receiver_id] pairs. Once this is
done, a flag is set for that event, so that the client can determine
if this has already been done or not.

Before the randomization, the previous results are purged, as they
have no use for that event any longer.